Fix display issue on Chrome on Android.
Chrome on Android has a known bug, see:
https://issuetracker.google.com/issues/36908439

This causes a background image with a `fixed` position that is not set
on the root or `<body>` element to scroll with the element contents
rather than be fixed relative to the browser viewport.

A related issue in the Chrome browser (desktop and mobile) causes
`mix-blend-mode` not to be applied to ancestor elements, but only to the
parent element of the element to which it is applied.

Taken together, this means that the commit below will prevent Chrome
from having the `difference` filter (`mix-blend-mode` value) show up on
either desktop or mobile, but at least it resolves the much worse
display issue with either the double-background image or an otherwise
unreadable main page content.

The same is true for Safari.

Firefox is unaffected.